tinystdio: Add stderr as separate buffer
And split posixiob into three files so that applications
need only include the ones they use.
